pyproject.toml
[project]
name = "google-cloud-ai-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.30.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
